backend/requirements.txt
python-dotenv==1.1.1
uvicorn==0.38.0
fastapi==0.115.6
pydantic==2.12.3
google-genai==1.45.0
psycopg2==2.9.11
python-multipart==0.0.20
minio==7.2.7
PyAudio==0.2.14